// src/types.ts
export interface InitialMovie {
  adult: boolean;
  backdrop_path: string;
  genre_ids?: number[];
  genres?: IGenre[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video?: boolean;
  vote_average: number;
  vote_count: number;
}

export interface IMovie {
  id: number;
  overview: string;
  title: string;
  poster: string;
  backdrop: string;
  genres: string[];
  releaseDate: string;
  rating: number;
}

export interface IGenre {
  id: number;
  name: string;
}